Update type of `nodeIndex` variable to avoid type conversions
Fix Clang warning `-Wimplicit-int-conversion`.

Fix MSVC warning:
> warning C4244: 'argument': conversion from 'int' to 'OP2Utility::Archive::AdaptiveHuffmanTree::NodeIndex', possible loss of data

Error was diagnosed at line 205, when `nodeIndex` was first used in a call to `IsLeaf`.

+++ src/Archive/HuffLZ.cpp
@@ -197,7 +197,7 @@ namespace OP2Utility::Archive
 	// Performs a tree lookup guided by the bitstream to find the next code
 	unsigned short HuffLZ::GetNextCode()
 	{
-		int nodeIndex;
+		unsigned short nodeIndex;
 		bool bBit;
 
 		// Use bitstream to find a terminal node
